+++ ios/Classes/SwiftScgatewayFlutterPlugin.swift
@@ -543,18 +543,56 @@ public class SwiftScgatewayFlutterPlugin: NSObject, FlutterPlugin {
      //MARK:- Launch smallplug
      else if(call.method == "launchSmallplug") {
         
-        SCGateway.shared.launchSmallPlug(presentingController: currentViewController, completion: {
-            response, error in
+        if let args = call.arguments as? Dictionary<String, Any>,
+           let smallplugHeaderText = args["smallplugHeaderText"] as? String {
+                
+            SCGateway.shared.launchSmallPlug(presentingController: currentViewController, smallplugHeaderText: smallplugHeaderText, completion: {
+                response, error in
+                
+                if let smallplugResponse = response {
+                    
+                    result(self.getJsonStringResult(success: true, data: smallplugResponse as? String, errorCode: nil, errorMessage: nil, transaction: nil))
+                    
+                }
+            })
             
-            if let smallplugResponse = response {
+        } else {
+            
+            SCGateway.shared.launchSmallPlug(presentingController: currentViewController, smallplugHeaderText: nil, completion: {
+                response, error in
                 
-                result(self.getJsonStringResult(success: true, data: smallplugResponse as? String, errorCode: nil, errorMessage: nil, transaction: nil))
-                
-            }
-        })
+                if let smallplugResponse = response {
+                    
+                    result(self.getJsonStringResult(success: true, data: smallplugResponse as? String, errorCode: nil, errorMessage: nil, transaction: nil))
+                    
+                }
+            })
+            
+        }
      }
     
-    
+     //MARK:- Launch smallplug with target
+     else if(call.method == "launchSmallplugWithTargetEndpoint") {
+        
+        if let args = call.arguments as? Dictionary<String, Any>,
+           
+           let target = args["targetEndpoint"] as? String,
+           let smallplugHeaderText = args["smallplugHeaderText"] as? String,
+           let params = args["params"] as? Dictionary<String, String> {
+            
+            SCGateway.shared.launchSmallPlug(presentingController: currentViewController, smallplugHeaderText: smallplugHeaderText, targetEndpoint: target, smallplugUrlParams: params, completion: {
+                response, error in
+                
+                if let smallplugResponse = response {
+                    
+                    result(self.getJsonStringResult(success: true, data: smallplugResponse as? String, errorCode: nil, errorMessage: nil, transaction: nil))
+                }
+            })
+            
+        }
+
+     }
+     
     else {
             result("Flutter method not implemented on iOS")
         }